+++ Operatoru/OperatorsLink.h
@@ -27,7 +27,7 @@ class OperatorsLink : public OperatorsBase {
     
     OperatorsLink(const string &aid1, const string &aid2, const string &nameAid1, const string &nameAid2);
     
-    void workAid(const int &, const double &);
+    void workAid(const int &, const int &);
     
     
 };
