+++ Operatoru/OperatorsPointers.h
@@ -27,7 +27,7 @@ public:
     
     OperatorsPointers(const string &aid1, const string &nameAid1, const string descAid1, const string &aid2,  const string &nameAid2, const string descAid2);
    
-    void workAid(const int &, const double &);
+    void workAid(const int &, const int &);
     
     
 };
